requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
python-docx==1.1.0
reportlab==4.0.7
python-pptx==0.6.23
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
googletrans==4.0.0rc1
Pillow==10.1.0
python-dotenv==1.0.0
